+++ examples/mouse_picking_2d.rs
@@ -1,36 +1,13 @@
 use bevy::{prelude::*, sprite::MaterialMesh2dBundle};
-use bevy_mod_raycast::{prelude::*, print_intersections};
+use bevy_mod_raycast::prelude::*;
 
 fn main() {
     App::new()
-        .add_plugins((
-            DefaultPlugins,
-            DefaultRaycastingPlugin::<MyRaycastSet>::default(),
-        ))
-        .add_systems(
-            First,
-            update_raycast_with_cursor.before(RaycastSystem::BuildRays::<MyRaycastSet>),
-        )
+        .add_plugins((DefaultPlugins, DeferredRaycastingPlugin::<()>::default()))
+        .insert_resource(RaycastPluginState::<()>::default().with_debug_cursor())
         .add_systems(Startup, setup)
-        .add_systems(Update, print_intersections::<MyRaycastSet>)
+        .add_systems(Update, print_intersections::<()>)
         .run();
-}
-
-#[derive(Reflect)]
-struct MyRaycastSet;
-
-// Update our `RaycastSource` with the current cursor position every frame.
-fn update_raycast_with_cursor(
-    mut cursor: EventReader<CursorMoved>,
-    mut query: Query<&mut RaycastSource<MyRaycastSet>>,
-) {
-    // Grab the most recent cursor event if it exists:
-    let Some(cursor_moved) = cursor.iter().last() else {
-        return;
-    };
-    for mut pick_source in &mut query {
-        pick_source.cast_method = RaycastMethod::Screenspace(cursor_moved.position);
-    }
 }
 
 fn setup(
@@ -38,15 +15,14 @@ fn setup(
     mut meshes: ResMut<Assets<Mesh>>,
     mut materials: ResMut<Assets<ColorMaterial>>,
 ) {
-    commands
-        .spawn(Camera2dBundle { ..default() })
-        .insert(RaycastSource::<MyRaycastSet>::new()); // Designate the camera as our source;
-    commands
-        .spawn(MaterialMesh2dBundle {
-            mesh: meshes.add(Mesh::from(shape::Quad::default())).into(),
+    commands.spawn((Camera2dBundle::default(), RaycastSource::<()>::new_cursor()));
+    commands.spawn((
+        MaterialMesh2dBundle {
+            mesh: meshes.add(Mesh::from(shape::Circle::default())).into(),
             transform: Transform::default().with_scale(Vec3::splat(128.)),
             material: materials.add(ColorMaterial::from(Color::PURPLE)),
             ..default()
-        })
-        .insert(RaycastMesh::<MyRaycastSet>::default()); // Make this mesh ray cast-able;
+        },
+        RaycastMesh::<()>::default(), // Make this mesh ray cast-able;
+    ));
 }