+++ examples/reflecting_laser.rs
@@ -1,12 +1,14 @@
 //! This example demonstrates how to use the [`Raycast`] system param to chain multiple raycasts and
 //! bounce off of surfaces.
 
-use bevy::{core_pipeline::bloom::BloomSettings, prelude::*};
+use std::f32::consts::{FRAC_PI_2, PI};
+
+use bevy::{core_pipeline::bloom::BloomSettings, math::vec3, prelude::*};
 use bevy_mod_raycast::prelude::*;
 
 fn main() {
     App::new()
-        .add_plugins(DefaultPlugins)
+        .add_plugins((DefaultPlugins, DefaultRaycastingPlugin))
         .add_systems(Startup, setup_scene)
         .add_systems(Update, bouncing_raycast)
         .insert_resource(ClearColor(Color::BLACK))
@@ -14,31 +16,48 @@ fn main() {
 }
 
 const MAX_BOUNCES: usize = 64;
-const LASER_MOVE_SPEED: f32 = 0.03;
+const LASER_SPEED: f32 = 0.03;
 
 #[derive(Reflect)]
 struct Laser;
 
-fn bouncing_raycast(mut raycast: Raycast, mut gizmos: Gizmos, time: Res<Time>) {
-    let t =
-        ((time.elapsed_seconds() - 4.0).max(0.0) * LASER_MOVE_SPEED).cos() * std::f32::consts::PI;
-    let mut ray_pos = Vec3::new(t.sin(), (3.0 * t).cos() * 0.5, t.cos()) * 0.5;
-    let mut ray_dir = (-ray_pos).normalize();
+fn bouncing_raycast(
+    mut raycast: Raycast,
+    mut gizmos: Gizmos,
+    time: Res<Time>,
+    cursor_ray: Res<CursorRay>,
+) {
+    let t = ((time.elapsed_seconds() - 4.0).max(0.0) * LASER_SPEED).cos() * std::f32::consts::PI;
+    let ray_pos = Vec3::new(t.sin(), (3.0 * t).cos() * 0.5, t.cos()) * 0.5;
+    let ray_dir = (-ray_pos).normalize();
+    let ray = Ray3d::new(ray_pos, ray_dir);
     gizmos.sphere(ray_pos, Quat::IDENTITY, 0.1, Color::WHITE);
+    bounce_ray(ray, &mut raycast, &mut gizmos, Color::RED);
 
+    if let Some(cursor_ray) = **cursor_ray {
+        bounce_ray(cursor_ray, &mut raycast, &mut gizmos, Color::GREEN)
+    }
+}
+
+fn bounce_ray(
+    mut ray: Ray3d,
+    raycast: &mut Raycast<'_, '_>,
+    gizmos: &mut Gizmos<'_>,
+    color: Color,
+) {
     let mut intersections = Vec::with_capacity(MAX_BOUNCES + 1);
-    intersections.push((ray_pos, Color::rgb(30.0, 0.0, 0.0)));
+    intersections.push((ray.origin(), Color::rgb(30.0, 0.0, 0.0)));
 
     for i in 0..MAX_BOUNCES {
-        let ray = Ray3d::new(ray_pos, ray_dir);
         if let Some((_, hit)) = raycast.cast_ray(ray, &RaycastSettings::default()).first() {
-            let r = 1.0 + 10.0 * (1.0 - i as f32 / MAX_BOUNCES as f32);
-            intersections.push((hit.position(), Color::rgb(r, 0.0, 0.0)));
-            gizmos.sphere(hit.position(), Quat::IDENTITY, 0.005, Color::RED * r * 2.0);
+            let bright = 1.0 + 10.0 * (1.0 - i as f32 / MAX_BOUNCES as f32);
+            intersections.push((hit.position(), color * bright));
+            gizmos.sphere(hit.position(), Quat::IDENTITY, 0.005, color * bright * 2.0);
+            let ray_dir = ray.direction();
             // reflect the ray
             let proj = (ray_dir.dot(hit.normal()) / hit.normal().dot(hit.normal())) * hit.normal();
-            ray_dir = (ray_dir - 2.0 * proj).normalize();
-            ray_pos = hit.position() + ray_dir * 1e-6;
+            ray.set_direction(ray_dir - 2.0 * proj);
+            ray.set_origin(hit.position() + ray.direction() * 1e-6);
         } else {
             break;
         }
@@ -54,7 +73,7 @@ fn setup_scene(
 ) {
     commands.spawn(DirectionalLightBundle {
         transform: Transform::from_rotation(Quat::from_euler(EulerRot::XYZ, -0.1, 0.2, 0.0)),
-        ..Default::default()
+        ..default()
     });
     commands.spawn((
         Camera3dBundle {
@@ -68,13 +87,21 @@ fn setup_scene(
         },
         BloomSettings::default(),
     ));
-    commands.spawn((
-        PbrBundle {
-            mesh: meshes.add(shape::Cube::default().into()),
-            material: materials.add(Color::GRAY.with_a(0.05).into()),
-            ..default()
-        },
-        // Without this, raycasts would shoot straight out from the inside of the cube.
-        bevy_mod_raycast::NoBackfaceCulling,
-    ));
+    // Make a box of planes facing inward so the laser gets trapped inside:
+    let plane = PbrBundle {
+        mesh: meshes.add(shape::Plane::default().into()),
+        material: materials.add(Color::GRAY.with_a(0.01).into()),
+        ..default()
+    };
+    let pbr_bundle = move |translation, rotation| PbrBundle {
+        transform: Transform::from_translation(translation)
+            .with_rotation(Quat::from_scaled_axis(rotation)),
+        ..plane.clone()
+    };
+    commands.spawn(pbr_bundle(vec3(0.0, 0.5, 0.0), Vec3::X * PI));
+    commands.spawn(pbr_bundle(vec3(0.0, -0.5, 0.0), Vec3::ZERO));
+    commands.spawn(pbr_bundle(vec3(0.5, 0.0, 0.0), Vec3::Z * FRAC_PI_2));
+    commands.spawn(pbr_bundle(vec3(-0.5, 0.0, 0.0), Vec3::Z * -FRAC_PI_2));
+    commands.spawn(pbr_bundle(vec3(0.0, 0.0, 0.5), Vec3::X * -FRAC_PI_2));
+    commands.spawn(pbr_bundle(vec3(0.0, 0.0, -0.5), Vec3::X * FRAC_PI_2));
 }